Adds crud and view mechanics for processing and rendering completed checklists.
